tape clean up 1.
This commit defines opr varin as extra arguments of the jvp and vjp.

Then the record only need to record the kwargs to the impl function
of the opr.

Add a rcd impl function to convert the kwargs to what actually shall
be recorded. This will be useful for linalg operators where only
a function that can create zeros_like(varin) is needed in vjp.